Converted geometry modules to use f64 instead of integers. Transcribed the ray tracing logic.

+++ src/geometry/vector.rs
@@ -1,16 +1,16 @@
 #[derive(Debug, Copy, Clone)]
 pub struct Vector{
-    pub x: i32,
-    pub y: i32,
-    pub z: i32
+    pub x: f64,
+    pub y: f64,
+    pub z: f64
 }
 
 impl Vector{
     pub fn origin() -> Vector {
-        Vector {x:0, y:0, z:0}
+        Vector {x:0.0, y:0.0, z:0.0}
     }
 
-    pub fn new(x: i32, y: i32, z: i32) -> Vector {
+    pub fn new(x: f64, y: f64, z: f64) -> Vector {
         Vector{x: x, y: y, z: z}
     }
 
@@ -22,11 +22,11 @@ impl Vector{
         Vector{x: self.x-other.x, y: self.y-other.y, z: self.z-other.z}
     }
 
-    pub fn dot(self, other: Vector) -> i32 {
+    pub fn dot(self, other: Vector) -> f64 {
         (self.x * other.x) + (self.y * other.y) + (self.z * other.z)
     }
 
     pub fn scale(self, factor: f64) -> Vector {
-        Vector{x: (self.x as f64 * factor) as i32, y: (self.y as f64 * factor) as i32, z: (self.z as f64 * factor) as i32}
+        Vector{x: (self.x * factor), y: (self.y * factor), z: (self.z * factor)}
     }
 }
